refactor: animation of single and multiple values
More explicit params and a single module for both functions
